Configuration class and main class shrink
